improve unit tests
