+++ rs_mcai_worker_sdk/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "mcai_worker_sdk"
-version = "0.11.10"
+version = "1.0.0-alpha1"
 authors = [
   "Marc-Antoine Arnaud <maarnaud@media-io.com>",
   "Valentin Noel <valentin.noel@media-io.com>"
